+++ src/renderer/src/components/terminal-pane/terminal-parser-handler-guard.ts
@@ -1,4 +1,4 @@
-import { recordRendererCrashBreadcrumb } from '@/lib/crash-diagnostics'
+import { recordRendererCrashBreadcrumb } from '@/lib/crash-breadcrumb-recorder'
 
 // Why: xterm's EscapeSequenceParser invokes custom CSI/OSC handlers
 // synchronously inside WriteBuffer._innerWrite, which has no try/catch. A
