+++ src/renderer/src/lib/pane-manager/pane-terminal-output-scheduler.test.ts
@@ -9,7 +9,7 @@ const mocks = vi.hoisted(() => ({
   recordRendererCrashBreadcrumb: vi.fn()
 }))
 
-vi.mock('@/lib/crash-diagnostics', () => ({
+vi.mock('@/lib/crash-breadcrumb-recorder', () => ({
   recordRendererCrashBreadcrumb: mocks.recordRendererCrashBreadcrumb
 }))
